requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.19
pytesseract==0.3.13
Pillow==11.0.0
requests==2.32.3

